refactor(report-calculations): extract shared payroll calculation utilities and isolate hourly validation logic
Why

- Shared payroll helpers were embedded in the hourly validation module, making reuse and testing harder.
- Hourly-only validation logic needed a clearer seam from generic calculation and date utilities.
- Report and debug consumers needed a single shared source for PAYE mode defaults and aggregation helpers.
- The extraction also preserves the current PAYE cumulative calculation path while simplifying module responsibilities.

What

- pwa/src/report/pay_calculations_shared.js Added a new shared utility module for rounding, tolerances, PAYE mode resolution, tax math, date helpers, catalog flag builders, and payment or deduction aggregation.
- pwa/src/report/pay_calculations_hourly.js Moved hourly-specific validation into its own module, kept pension and PAYE validation exports focused on hourly records, and switched imports to the shared calculation helpers.
- pwa/src/report/report_calculations.js Repointed report entry validation and shared sum helpers to the new module split.
- pwa/src/ui/debug_tools.js Switched PAYE cumulative default mode imports to the shared calculation module.
- tests/report_calculation.test.mjs Updated imports so shared helpers come from the shared module and hourly validation comes from the hourly-only module.
- tests/report_workflow.test.mjs Updated workflow tests to import shared aggregation helpers from the extracted shared module.

Impact

- Improves maintainability by separating worker-specific validation from reusable payroll calculation primitives.
- Makes future worker-type expansion easier because shared math and date logic no longer depend on hourly validation modules.
- Reduces coupling across report and debug code paths by centralizing common calculation utilities.
- Keeps regression coverage aligned with the new module boundaries so refactoring remains safer.
