+++ pwa/src/ui/debug_tools.js
@@ -1,6 +1,6 @@
 import { UNKNOWN_APP_VERSION } from './app_version.js'
 import { ACTIVE_PAYROLL_FORMAT } from '../parse/active_format.js'
-import { PAYE_CUMULATIVE_DEFAULT_MODE_BY_PAYROLL_FORMAT } from '../report/hourly_pay_calculations.js'
+import { PAYE_CUMULATIVE_DEFAULT_MODE_BY_PAYROLL_FORMAT } from '../report/pay_calculations_shared.js'
 
 /** @type {'exact' | 'sage_approx' | 'table_mode'} */
 const DEFAULT_PAYE_CUMULATIVE_MODE =
